pyproject.toml
[project]
name = "aganitha-paper-fetcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Shivangi", email = "shivangimaurya1007@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython (>=1.85,<2.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "rich (>=14.0.0,<15.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "aganitha" }
]

[tool.poetry.scripts]
get-papers-list = "cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

